Move future files out of internal/. (#1450)
The public headers for `future<T>` should be in google/cloud/ so they
can be picked up by the Doxygen scripts. This fixes #1345.
