+++ google/cloud/CMakeLists.txt
@@ -93,6 +93,8 @@ configure_file(internal/build_info.cc.in internal/build_info.cc)
 # the client library
 add_library(google_cloud_cpp_common
             future.h
+            future_generic.h
+            future_void.h
             iam_binding.h
             iam_bindings.h
             iam_bindings.cc
@@ -111,8 +113,6 @@ add_library(google_cloud_cpp_common
             internal/future_impl.cc
             internal/future_then_impl.h
             internal/future_then_meta.h
-            internal/future_generic.h
-            internal/future_void.h
             internal/getenv.h
             internal/getenv.cc
             internal/make_unique.h
@@ -171,15 +171,15 @@ target_link_libraries(google_cloud_cpp_testing
 create_bazel_config(google_cloud_cpp_testing)
 
 set(google_cloud_cpp_common_unit_tests
+    future_generic_test.cc
+    future_generic_then_test.cc
+    future_void_test.cc
+    future_void_then_test.cc
     iam_bindings_test.cc
     internal/backoff_policy_test.cc
     internal/big_endian_test.cc
     internal/filesystem_test.cc
     internal/future_impl_test.cc
-    internal/future_generic_test.cc
-    internal/future_generic_then_test.cc
-    internal/future_void_test.cc
-    internal/future_void_then_test.cc
     internal/invoke_result_test.cc
     internal/random_test.cc
     internal/retry_policy_test.cc
