+++ google/cloud/future_generic_then_test.cc
