+++ google/cloud/future_void_then_test.cc
