+++ google/cloud/google_cloud_cpp_common.bzl
@@ -1,6 +1,8 @@
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 google_cloud_cpp_common_HDRS = [
     "future.h",
+    "future_generic.h",
+    "future_void.h",
     "iam_binding.h",
     "iam_bindings.h",
     "iam_policy.h",
@@ -13,8 +15,6 @@ google_cloud_cpp_common_HDRS = [
     "internal/future_impl.h",
     "internal/future_then_impl.h",
     "internal/future_then_meta.h",
-    "internal/future_generic.h",
-    "internal/future_void.h",
     "internal/getenv.h",
     "internal/make_unique.h",
     "internal/port_platform.h",
