+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -1,14 +1,14 @@
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 google_cloud_cpp_common_unit_tests = [
+    "future_generic_test.cc",
+    "future_generic_then_test.cc",
+    "future_void_test.cc",
+    "future_void_then_test.cc",
     "iam_bindings_test.cc",
     "internal/backoff_policy_test.cc",
     "internal/big_endian_test.cc",
     "internal/filesystem_test.cc",
     "internal/future_impl_test.cc",
-    "internal/future_generic_test.cc",
-    "internal/future_generic_then_test.cc",
-    "internal/future_void_test.cc",
-    "internal/future_void_then_test.cc",
     "internal/invoke_result_test.cc",
     "internal/random_test.cc",
     "internal/retry_policy_test.cc",
